+++ src/higherOrderComponents/withAuthUser.tsx
@@ -20,6 +20,8 @@
 import * as React from 'react';
 import {AuthContextConsumer} from '../AuthContext';
 import {AuthStateUserObject} from '../types';
+import {AuthKitError} from '../errors';
+import {isAuthenticated} from '../utils/utils';
 
 /**
  * @interface withAuthProps
@@ -43,10 +45,10 @@ function withAuthUser<P extends withAuthProps>(
         {(context) => {
           if (context === null) {
             throw new
-            Error('Auth Provider is missing. ' +
+            AuthKitError('Auth Provider is missing. ' +
               'Please add the AuthProvider before Router');
           }
-          if (context.authState.auth) {
+          if (context.authState.auth && isAuthenticated(context.authState)) {
             return (
               <Component {...props} authState={context.authState.userState}/>
             );