+++ src/higherOrderComponents/withIsAuthenticated.tsx
@@ -21,6 +21,8 @@
 import * as React from 'react';
 import {AuthContextConsumer} from '../AuthContext';
 import {doSignOut} from '../utils/reducers';
+import {AuthKitError} from '../errors';
+import {isAuthenticated} from '../utils/utils';
 
 /**
  * @interface withAuthHeaderProps
@@ -45,17 +47,13 @@ function withIsAuthenticated<P extends withAuthHeaderProps>(
         {(c) => {
           if (c === null) {
             throw new
-            Error('Auth Provider is missing. ' +
+            AuthKitError('Auth Provider is missing. ' +
               'Please add the AuthProvider before Router');
           }
-          if (c.authState.auth) {
-            if (new Date(c.authState.auth.expiresAt) > new Date()) {
-              return <Component {...props} isAuth={true}/>;
-            } else {
-              c.dispatch(doSignOut());
-              return <Component {...props} isAuth={false}/>;
-            }
+          if (c.authState.auth && isAuthenticated(c.authState)) {
+            return <Component {...props} isAuth={true}/>;
           } else {
+            c.dispatch(doSignOut());
             return <Component {...props} isAuth={false}/>;
           }
         }}