+++ src/hooks/useAuthHeader.ts
@@ -20,6 +20,8 @@
 
 import * as React from 'react';
 import AuthContext from '../AuthContext';
+import {AuthKitError} from '../errors';
+import {isAuthenticated} from '../utils/utils';
 
 /**
  *
@@ -28,13 +30,13 @@ function useAuthHeader(): () => (string) {
   const c = React.useContext(AuthContext);
   if (c === null) {
     throw new
-    Error('Auth Provider is missing. ' +
+    AuthKitError('Auth Provider is missing. ' +
       'Please add the AuthProvider before Router');
   }
 
 
   return () => {
-    if (c.authState.auth) {
+    if (c.authState.auth && isAuthenticated(c.authState)) {
       return `${c.authState.auth.type} ${c.authState.auth.token}`;
     } else {
       return ``;